test: mirror test layout to src/ package structure
